加入GitHub相关CI文件,但是并没有实现.调整app中部分内容实现方式.
