+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,36 @@
+---
+name: '报告Bug 🐛'
+about: 报告 Destiny Flow 的 bug
+title: '🐛[BUG]'
+labels: '🐛bug'
+assignees: ''
+---
+
+### 🐛 bug 描述
+
+<!--
+详细地描述 bug，让大家都能理解
+-->
+
+### 📷 复现步骤
+
+<!--
+清晰描述复现步骤，让别人也能看到问题
+-->
+
+### 🏞 期望结果
+
+<!--
+描述你原本期望看到的结果
+-->
+
+### © 版本信息
+
+- 浏览器环境
+- 操作系统 [e.g. mac OS]
+
+### 🚑 其他信息
+
+<!--
+如截图等其他信息可以贴在这里
+-->
