+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,19 @@
+---
+name: '疑问或需要帮助 ❓'
+about: 对 Destiny Flow 使用的疑问或需要帮助
+title: '🧐[问题]'
+labels: '🧐question'
+assignees: ''
+---
+
+### 🧐 问题描述
+
+<!--
+详细地描述问题，让大家都能理解
+-->
+
+### 🚑 其他信息
+
+<!--
+如截图等其他信息可以贴在这里
+-->
