+++ .github/workflows/rebase.yml
@@ -1,0 +1,17 @@
+# on:
+#   issue_comment:
+#     types: [created]
+# name: Automatic Rebase
+# jobs:
+#   rebase:
+#     name: Rebase
+#     if: github.event.issue.pull_request != '' && contains(github.event.comment.body, '/rebase')
+#     runs-on: ubuntu-latest
+#     steps:
+#       - uses: actions/checkout@master
+#         with:
+#           fetch-depth: 0
+#       - name: Automatic Rebase
+#         uses: cirrus-actions/rebase@1.2
+#         env:
+#           GITHUB_TOKEN: ${{ secrets.ACTION_TOKEN }}
