+++ .vscode/settings.json
@@ -24,6 +24,7 @@
   "workbench.iconTheme": "material-icon-theme",
   "javascript.updateImportsOnFileMove.enabled": "always",
   "window.zoomLevel": 0,
+  "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
   "cSpell.words": [
     "APIDATA",
     "AUTOGENERATED",
@@ -191,6 +192,7 @@
     "projectview",
     "psize",
     "pwacompat",
+    "qixian",
     "qyrmbhuo",
     "regiondata",
     "reportmanage",
