+++ config/default.ts
@@ -2,24 +2,18 @@
 
 import { Settings as ProSettings } from '@ant-design/pro-layout';
 
-type DefaultSettings = ProSettings & {
-  pwa: boolean;
-};
-
-const proSettings: DefaultSettings = {
+export default {
   navTheme: 'light',
-  primaryColor: '#F5222D',
+  // 拂晓蓝
+  primaryColor: '#1890ff',
   layout: 'mix',
   contentWidth: 'Fluid',
   fixedHeader: false,
+  autoHideHeader: false,
   fixSiderbar: false,
   colorWeak: false,
   menu: { locale: true },
   title: 'Destiny Flow',
   pwa: false,
   iconfontUrl: ''
-};
-
-export type { DefaultSettings };
-
-export default proSettings;
+} as ProSettings;
