+++ src/global.less
@@ -21,6 +21,12 @@ body,
   border-radius: 7px !important;
 }
 
+.ant-pro-setting-drawer-handle {
+  width: 32px !important;
+  height: 32px !important;
+  top: 300px !important;
+}
+
 canvas {
   display: block;
 }
